MAINT: remove erfcinv
