+++ scipy/special/functions.json
@@ -162,11 +162,6 @@
             "erfinv_double": "d->d"
         }
     },
-    "erfcinv": {
-        "xsf_wrappers.h": {
-            "cephes_erfcinv": "d->d"
-        }
-    },
     "eval_chebyc": {
         "orthogonal_eval.pxd": {
             "eval_chebyc[double complex]": "dD->D",
